Move Duration to config
