--- pkg/sqlutil/duration.go
@@ -1,123 +1,0 @@
-package sqlutil
-
-import (
-	"database/sql/driver"
-	"encoding/json"
-	"fmt"
-	"time"
-
-	"github.com/pkg/errors"
-)
-
-// Duration is a non-negative time duration.
-type Duration struct{ d time.Duration }
-
-func MakeDuration(d time.Duration) (Duration, error) {
-	if d < time.Duration(0) {
-		return Duration{}, fmt.Errorf("cannot make negative time duration: %s", d)
-	}
-	return Duration{d: d}, nil
-}
-
-func ParseDuration(s string) (Duration, error) {
-	d, err := time.ParseDuration(s)
-	if err != nil {
-		return Duration{}, err
-	}
-
-	return MakeDuration(d)
-}
-
-func MustMakeDuration(d time.Duration) Duration {
-	rv, err := MakeDuration(d)
-	if err != nil {
-		panic(err)
-	}
-	return rv
-}
-
-func MustNewDuration(d time.Duration) *Duration {
-	rv := MustMakeDuration(d)
-	return &rv
-}
-
-// Duration returns the value as the standard time.Duration value.
-func (d Duration) Duration() time.Duration {
-	return d.d
-}
-
-// Before returns the time d units before time t
-func (d Duration) Before(t time.Time) time.Time {
-	return t.Add(-d.Duration())
-}
-
-// Shorter returns true if and only if d is shorter than od.
-func (d Duration) Shorter(od Duration) bool { return d.d < od.d }
-
-// IsInstant is true if and only if d is of duration 0
-func (d Duration) IsInstant() bool { return d.d == 0 }
-
-// String returns a string representing the duration in the form "72h3m0.5s".
-// Leading zero units are omitted. As a special case, durations less than one
-// second format use a smaller unit (milli-, micro-, or nanoseconds) to ensure
-// that the leading digit is non-zero. The zero duration formats as 0s.
-func (d Duration) String() string {
-	return d.Duration().String()
-}
-
-// MarshalJSON implements the json.Marshaler interface.
-func (d Duration) MarshalJSON() ([]byte, error) {
-	return json.Marshal(d.String())
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface.
-func (d *Duration) UnmarshalJSON(input []byte) error {
-	var txt string
-	err := json.Unmarshal(input, &txt)
-	if err != nil {
-		return err
-	}
-	v, err := time.ParseDuration(txt)
-	if err != nil {
-		return err
-	}
-	*d, err = MakeDuration(v)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (d *Duration) Scan(v interface{}) (err error) {
-	switch tv := v.(type) {
-	case int64:
-		*d, err = MakeDuration(time.Duration(tv))
-		return err
-	default:
-		return errors.Errorf(`don't know how to parse "%s" of type %T as a `+
-			`models.Duration`, tv, tv)
-	}
-}
-
-func (d Duration) Value() (driver.Value, error) {
-	return int64(d.d), nil
-}
-
-// MarshalText implements the text.Marshaler interface.
-func (d Duration) MarshalText() ([]byte, error) {
-	return []byte(d.d.String()), nil
-}
-
-// UnmarshalText implements the text.Unmarshaler interface.
-func (d *Duration) UnmarshalText(input []byte) error {
-	v, err := time.ParseDuration(string(input))
-	if err != nil {
-		return err
-	}
-	pd, err := MakeDuration(v)
-	if err != nil {
-		return err
-	}
-	*d = pd
-	return nil
-}
